Fix VAD false positive spam causing continuous [BLANK_AUDIO] detections
Fixed two critical bugs in Voice Activity Detection:

1. Type mismatch bug in talk-llama.cpp:
   - vad_last_ms was declared as float but vad_simple() expects int
   - This caused memory corruption, turning 1250ms into garbage values
     like 1404127495ms
   - With such huge values, buffer size checks always failed
   - Fix: Changed vad_last_ms from float to int (line 146)

2. Inverted logic bug in common.cpp vad_simple():
   - When audio buffer was insufficient (< 1250ms samples), the function
     returned false (speech detected) instead of true (silence)
   - During startup and after each transcription, buffer gradually fills
     from 0 to 20480 samples
   - While buffer < 20000 samples, VAD would trigger continuously on
     background noise
   - This caused endless [BLANK_AUDIO] loop with "Heard nothing, skipping"
   - Fix: Changed return value from false to true when buffer insufficient
     (line 769)

Result: VAD now correctly waits for actual speech above min_energy
threshold (0.0012) without false triggers on background noise or during
buffer fill-up.

+++ custom/talk-llama/talk-llama.cpp
@@ -143,7 +143,7 @@ struct whisper_params
 
 	float vad_thold = 0.6f;
 	float vad_start_thold = 0.000270f; // 0 to turn off, you can see your current energy_last (loudness level) when running with --print-energy param
-	float vad_last_ms = 1250;
+	int vad_last_ms = 1250;  // Changed from float to int to match vad_simple signature
 	float freq_thold = 100.0f;
 	float min_energy = 0.0012f; // Minimum energy threshold to prevent TTS feedback
